Use POH for pinned data

+++ ManagedCLR.JIT/BaseJIT.cs
@@ -2,6 +2,6 @@
 {
 	public abstract class BaseJIT
 	{
-		public abstract void WriteStub(uint slot, nuint memory, int maxStack);
+		public abstract void WriteStub(nuint slot, nuint memory, int maxStack);
 	}
 }
